chore: move normalizeEndpoint to own file
