+++ packages/config-resolver/src/endpointsConfig/normalizeEndpoint.ts
@@ -1,0 +1,15 @@
+import { Endpoint, Provider } from "@aws-sdk/types";
+
+import { EndpointsInputConfig, PreviouslyResolved } from "./configurations";
+
+export const normalizeEndpoint = (input: EndpointsInputConfig & PreviouslyResolved): Provider<Endpoint> => {
+  const { endpoint, urlParser } = input;
+  if (typeof endpoint === "string") {
+    const promisified = Promise.resolve(urlParser(endpoint));
+    return () => promisified;
+  } else if (typeof endpoint === "object") {
+    const promisified = Promise.resolve(endpoint);
+    return () => promisified;
+  }
+  return endpoint!;
+};
